refactor(nodesresolver): change nodesresolver to nodesnetworkresolver
The Nodes Resolver is currently not needed this is why we repurposed the resolver to be a
nodesNetwork resolver.
